Renamed spec to tests, and split validation into results and validation modules
