+++ mlte/validation/validator.py
@@ -15,8 +15,8 @@ from mlte._private.fixed_json import json
 from mlte._private.function_info import FunctionInfo
 from mlte.model.base_model import BaseModel
 from mlte.model.serializable import Serializable
-from mlte.validation.model_validator import ValidatorModel
-from mlte.validation.result import Failure, Info, Result, Success
+from mlte.results.result import Failure, Info, Result, Success
+from mlte.validation.model import ValidatorModel
 
 
 class Validator(Serializable):
