+++ test/fixture/artifact.py
@@ -39,8 +39,8 @@ from mlte.report.model import (
     QuantitiveAnalysisDescriptor,
     ReportModel,
 )
-from mlte.spec.model import TestCaseModel, TestSuiteModel
-from mlte.validation.model import ResultModel, TestResultsModel
+from mlte.results.model import ResultModel, TestResultsModel
+from mlte.tests.model import TestCaseModel, TestSuiteModel
 from mlte.validation.validator import Validator
 from test.evidence.types.helper import get_sample_evidence_metadata
