+++ test/results/test_test_results.py
@@ -13,10 +13,10 @@ import pytest
 from mlte.artifact.type import ArtifactType
 from mlte.context.context import Context
 from mlte.evidence.types.integer import Integer
-from mlte.spec.test_suite import TestSuite
+from mlte.results.result import Result
+from mlte.results.test_results import TestResults
 from mlte.store.artifact.store import ArtifactStore
-from mlte.validation.result import Result
-from mlte.validation.test_results import TestResults
+from mlte.tests.test_suite import TestSuite
 from mlte.validation.test_suite_validator import TestSuiteValidator
 from test.evidence.types.helper import get_sample_evidence_metadata
 from test.fixture.artifact import ArtifactFactory
