+++ test/tests/test_case.py
@@ -5,8 +5,8 @@ Unit tests for TestCase functionality.
 from __future__ import annotations
 
 from mlte.evidence.types.integer import Integer
-from mlte.spec.test_case import TestCase
-from mlte.validation.result import Failure, Success
+from mlte.results.result import Failure, Success
+from mlte.tests.test_case import TestCase
 from mlte.validation.validator import Validator
 from test.evidence.types.helper import get_sample_evidence_metadata
