+++ tools/schema.py
@@ -82,14 +82,12 @@ CONFIGS = [
         sub_path="artifact",
     ),
     Config(
-        model=ModelImport(path="mlte.spec.model", model="TestSuiteModel"),
+        model=ModelImport(path="mlte.tests.model", model="TestSuiteModel"),
         output_path="spec/v0.0.1",
         sub_path="artifact",
     ),
     Config(
-        model=ModelImport(
-            path="mlte.validation.model", model="TestResultsModel"
-        ),
+        model=ModelImport(path="mlte.results.model", model="TestResultsModel"),
         output_path="validated/v0.0.1",
         sub_path="artifact",
     ),
